Use decimal fields for article combo data

+++ src/ERPAccounting.Application/DTOs/ComboDtos.cs
@@ -59,11 +59,11 @@ public record ArticleComboDto(
     string NazivArtikla,
     string? JedinicaMere,
     string? IdPoreskaStopa,
-    double ProcenatPoreza,
-    double Akciza,
-    double KoeficijentKolicine,
+    decimal ProcenatPoreza,
+    decimal Akciza,
+    decimal KoeficijentKolicine,
     bool ImaLot,
-    double? OtkupnaCena,
+    decimal? OtkupnaCena,
     bool PoljoprivredniProizvod
 );
